+++ Assignment.Interfaces/Assignment.Interfaces.csproj
@@ -1,0 +1,5 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <RootNamespace>Assignment</RootNamespace>
+  </PropertyGroup>
+</Project>
