--- Assignment/Assignment/People.csv
@@ -1,51 +1,0 @@
-Id,FirstName,LastName,Email,StreetAddress,City,State,Zip
-1,Priscilla,Jenyns,pjenyns0@state.gov,7884 Corry Way,Helena,MT,70577
-2,Karin,Joder,kjoder1@quantcast.com,03594 Florence Park,Tampa,FL,71961
-3,Chadd,Stennine,cstennine2@wired.com,94148 Kings Terrace,Long Beach,CA,59721
-4,Fremont,Pallaske,fpallaske3@umich.edu,16958 Forster Crossing,Atlanta,GA,10687
-5,Melisa,Kerslake,mkerslake4@dion.ne.jp,283 Pawling Parkway,Dallas,TX,88632
-6,Darline,Brauner,dbrauner5@biglobe.ne.jp,33 Menomonie Trail,Atlanta,GA,10687
-7,Issiah,Bester,ibester6@psu.edu,96 Independence Hill,Richmond,VA,37731
-8,Joly,Scneider,jscneider7@pagesperso-orange.fr,53 Grim Point,Spokane,WA,99022
-9,Swen,Mulvany,smulvany8@surveymonkey.com,71258 Cambridge Avenue,Erie,PA,53141
-10,Scarface,Dennington,sdennington9@chron.com,9978 Maple Wood Parkway,Memphis,TN,99928
-11,Henri,Dorr,hdorra@exblog.jp,2646 Hazelcrest Road,San Francisco,CA,40486
-12,Vince,Dee,vdeeb@japanpost.jp,98027 Ridgeview Lane,Houston,TX,67224
-13,Daile,Croser,dcroserc@reverbnation.com,3699 Duke Parkway,Winston Salem,NC,44527
-14,Editha,Loseke,eloseked@posterous.com,39431 Lotheville Pass,Washington,DC,73484
-15,Phillida,Chastagnier,pchastagniere@reference.com,1 Rutledge Point,Spokane,WA,99021
-16,Ewart,Puckinghorne,epuckinghornef@indiatimes.com,9 Forster Lane,Lincoln,NE,40053
-17,Sancho,Mahony,smahonyg@stanford.edu,90 Birchwood Street,Las Vegas,NV,36230
-18,Iggy,Baughen,ibaughenh@addthis.com,95574 Pond Crossing,Winston Salem,NC,76871
-19,Fayette,Dougherty,fdoughertyi@stanford.edu,6487 Pepper Wood Court,Spokane,WA,99021
-20,Chelsy,Buckle,cbucklej@tiny.cc,38 Talisman Hill,Jacksonville,FL,57619
-21,Reade,Stalman,rstalmank@purevolume.com,20622 Fulton Street,Atlanta,GA,10687
-22,Amelia,Toal,atoall@fema.gov,4667 Jay Plaza,Huntington,WV,44302
-23,Jermaine,Danelutti,jdaneluttim@jimdo.com,7 Onsgard Lane,Frederick,MD,56551
-24,Karel,Buttery,kbutteryn@woothemes.com,1288 Petterle Trail,Charlotte,NC,39115
-25,Jedd,Boissier,jboissiero@cbsnews.com,1 Arrowood Crossing,San Diego,CA,96101
-26,Molly,Jeannot,mjeannotp@google.ca,00546 International Alley,Tucson,AZ,94123
-27,Maria,Rawsthorne,mrawsthorneq@slate.com,609 Kedzie Alley,New York City,NY,46123
-28,Mayor,Thurnham,mthurnhamr@live.com,2 Eastlawn Center,Las Vegas,NV,87234
-29,Dorree,Luke,dlukes@microsoft.com,3 Tomscot Parkway,Wichita,KS,43581
-30,Stern,Aucutt,saucuttt@live.com,979 Magdeline Way,Washington,DC,61975
-31,Ev,Challace,echallaceu@nasa.gov,577 Arkansas Parkway,San Diego,CA,92805
-32,Bernette,Felstead,bfelsteadv@ycombinator.com,18 Center Point,San Diego,CA,69152
-33,Bone,Spittall,bspittallw@delicious.com,18118 South Avenue,New Orleans,LA,74634
-34,Sayres,Rumble,srumblex@addthis.com,56 Milwaukee Park,Minneapolis,MN,17309
-35,Marlow,Gossart,mgossarty@elpais.com,711 Cambridge Court,Pasadena,TX,73914
-36,Adrea,Lay,alayz@spotify.com,79 Pond Park,Salt Lake City,UT,63481
-37,Westley,Mesnard,wmesnard10@amazonaws.com,075 Pierstorff Road,Manchester,NH,66946
-38,Gabrielle,Vitler,gvitler11@comcast.net,58 Texas Alley,Eugene,OR,17222
-39,Celestyna,Robken,crobken12@t.co,27 Moland Parkway,San Antonio,TX,74488
-40,Selia,Bowe,sbowe13@360.cn,42 Lerdahl Plaza,New Orleans,LA,13260
-41,Kellen,Tawn,ktawn14@istockphoto.com,463 International Crossing,San Antonio,TX,19079
-42,Kordula,Whellams,kwhellams15@ebay.com,19 Melvin Trail,Brooklyn,NY,13751
-43,Joellen,Mullany,jmullany16@eepurl.com,2721 Morningstar Street,Lees Summit,MO,63478
-44,Audrye,Reolfo,areolfo17@51.la,94 Bayside Alley,Philadelphia,PA,15141
-45,Horace,Han,hhan18@nifty.com,33737 Waubesa Center,Spartanburg,SC,98413
-46,Marijn,McKennan,mmckennan19@usa.gov,4 Leroy Center,Washington,DC,43177
-47,Buck,Aspin,baspin1a@myspace.com,286 Springs Circle,Terre Haute,IN,72565
-48,Constantia,Sogg,csogg1b@list-manage.com,6217 Amoth Road,Gaithersburg,MD,79967
-49,Arthur,Myles,amyles1c@miibeian.gov.cn,4718 Thackeray Pass,Mobile,AL,37308
-50,Claudell,Leathe,cleathe1d@columbia.edu,30262 Steensland Way,Newport News,VA,87930
